restructure and renaming
