--- components/navbar/navbar.module.css
@@ -1,17 +1,0 @@
-
-@media only screen 
-  and (min-device-width: 375px) 
-  and (max-device-width: 812px) 
-  and (-webkit-min-device-pixel-ratio: 3) { 
-
-.Navbar{
-
-    border:1px solid black;
-    display:grid;
-    grid-template-columns: 1fr 1fr;
-}
-
-
-
-
-}
