+++ pages/home/index.js
@@ -1,15 +1,14 @@
 import React from "react";
 import "antd/dist/antd.css";
-import SearchBar from "../../components/Searchbar";
-
+import SearchBar from "/components/SearchBar/SearchBar";
 
 const SearchPage = () => {
-  return (
-    <div>
-      <h1>Search Page</h1>
-      <SearchBar />
-    </div>
-  );
+	return (
+		<div>
+			<h1>Search Page</h1>
+			<SearchBar />
+		</div>
+	);
 };
 
 export default SearchPage;
