+++ styles/globals.css
@@ -1,107 +1,89 @@
 html,
 body {
-  padding: 0;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+	padding: 0;
+	margin: 0;
+	font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
+		Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
 
+/* [START] THIS IS STYLING FOR THE HEADER - DO NOT CHANGE */
+.header {
+	display: flex;
+	width: 100%;
+	align-items: center;
+	justify-content: space-between;
+	position: a;
+	padding: 0.4rem;
+	background: orange;
+}
 
+.list_items {
+	list-style-type: none;
+	display: flex;
+	gap: 3.5rem;
+}
 
-.header{
-  display: flex;
-  width:100%;
-  align-items: center;
-  justify-content: space-between;
-  position: a;
-  padding:.4rem;
-  background:orange;
-  }
-  
-  
-  .list_items{
-  list-style-type: none;
-  display: flex;
-  gap:3.5rem;
-     
-  }
-  
-  
-  
-  .list_items .list a{
-  
-      text-decoration: none;
-      color: white;
-      font-size: clamp(1rem, .5vw + .5rem, 3.5rem );
-  }
-  
-  
-  
-  
-  .list_items .list a:hover, .list_items .list a:focus{
-  
-      color:navy
-  }
-  
-  .close_icon{
-      display: none;
-  }
-  
-  .hamburger_icon .icon{
-      color: navy;
-      font-size:clamp(1rem,4vw + 1rem ,4rem) ;   
-      cursor: pointer;
-      display:none;
-  }
-  
-  @media(max-width:50em)  { 
-       .list_items{
-    
-          width:46%;
-          padding: 10px;
-          flex-direction: column;
-          position: fixed;
-          align-items: center;
-          justify-content: center;
-          gap:4rem;
-          background-color:rgba(255, 166, 0, 0.39);
-          backdrop-filter: blur(1rem);
-        transform:translateX(110%);
-        transition: 500ms;
-          inset: 0,0,0,30;
-          border-radius: 15px;
-       } 
-  
-       .list_items.active{
-          
-  transform:translateX(0) ;
-  
-       }
-  
-     .list_items li a {
-  font-size: clamp(.1rem,1vw + .5rem ,3.5rem);
-  
-  
-     } 
-     .list_items .list a{
-      color:black;
-  
-     }
-  .close_icon{
-  display:block ;
-  position: absolute;
-  left:2%;
-  top: 2%;
-  cursor: pointer;
-  color: rgb(58, 55, 55);
-  }
-  
-  
-  .hamburger_icon .icon{
-  
-      display:block;
-  }
-  
-  }
+.list_items .list a {
+	text-decoration: none;
+	color: white;
+	font-size: clamp(1rem, 0.5vw + 0.5rem, 3.5rem);
+}
 
+.list_items .list a:hover,
+.list_items .list a:focus {
+	color: navy;
+}
 
+.close_icon {
+	display: none;
+}
+
+.hamburger_icon .icon {
+	color: navy;
+	font-size: clamp(1rem, 4vw + 1rem, 4rem);
+	cursor: pointer;
+	display: none;
+}
+
+@media (max-width: 50em) {
+	.list_items {
+		width: 46%;
+		padding: 10px;
+		flex-direction: column;
+		position: fixed;
+		align-items: center;
+		justify-content: center;
+		gap: 4rem;
+		background-color: rgba(255, 166, 0, 0.39);
+		backdrop-filter: blur(1rem);
+		transform: translateX(110%);
+		transition: 500ms;
+		inset: 0, 0, 0, 30;
+		border-radius: 15px;
+	}
+
+	.list_items.active {
+		transform: translateX(0);
+	}
+
+	.list_items li a {
+		font-size: clamp(0.1rem, 1vw + 0.5rem, 3.5rem);
+	}
+	.list_items .list a {
+		color: black;
+	}
+	.close_icon {
+		display: block;
+		position: absolute;
+		left: 2%;
+		top: 2%;
+		cursor: pointer;
+		color: rgb(58, 55, 55);
+	}
+
+	.hamburger_icon .icon {
+		display: block;
+	}
+}
+
+/* [END] THIS IS STYLING FOR THE HEADER - DO NOT CHANGE */
